Added check for null pointer

+++ aFWall/src/main/java/dev/ukanth/ufirewall/service/ToggleTileService.java
@@ -35,16 +35,18 @@ public class ToggleTileService extends TileService {
         super.onStartListening();
         boolean status = Api.isEnabled(this);
         Tile tile = getQsTile(); // this is getQsTile() method form java, used in Kotlin as a property
-        if (!status) {
-            tile.setLabel(getString(R.string.inactive));
-            tile.setIcon(Icon.createWithResource(this, R.drawable.notification_error));
-            tile.setState(Tile.STATE_INACTIVE);
-        } else {
-            tile.setLabel(getString(R.string.active));
-            tile.setIcon(Icon.createWithResource(this, R.drawable.notification));
-            tile.setState(Tile.STATE_ACTIVE);
+        if (tile != null) {
+            if (!status) {
+                tile.setLabel(getString(R.string.inactive));
+                tile.setIcon(Icon.createWithResource(this, R.drawable.notification_error));
+                tile.setState(Tile.STATE_INACTIVE);
+            } else {
+                tile.setLabel(getString(R.string.active));
+                tile.setIcon(Icon.createWithResource(this, R.drawable.notification));
+                tile.setState(Tile.STATE_ACTIVE);
+            }
+            tile.updateTile();
         }
-        tile.updateTile();
     }
 
     @Override
@@ -69,36 +71,38 @@ public class ToggleTileService extends TileService {
 
         Tile tile = getQsTile();
 
-        if (enabled) {
-            Api.applySavedIptablesRules(context, true, new RootCommand()
-                    .setSuccessToast(R.string.toast_enabled)
-                    .setFailureToast(R.string.toast_error_enabling)
-                    .setReopenShell(true)
-                    .setCallback(new RootCommand.Callback() {
-                        public void cbFunc(RootCommand state) {
-                            // setEnabled always sends us a STATUS_CHANGED_MSG intent to update the icon
-                            Api.setEnabled(context, state.exitCode == 0, true);
+        if (tile != null) {
+            if (enabled) {
+                Api.applySavedIptablesRules(context, true, new RootCommand()
+                        .setSuccessToast(R.string.toast_enabled)
+                        .setFailureToast(R.string.toast_error_enabling)
+                        .setReopenShell(true)
+                        .setCallback(new RootCommand.Callback() {
+                            public void cbFunc(RootCommand state) {
+                                // setEnabled always sends us a STATUS_CHANGED_MSG intent to update the icon
+                                Api.setEnabled(context, state.exitCode == 0, true);
 
-                            tile.setState(Tile.STATE_ACTIVE);
-                            tile.setLabel(getString(R.string.active));
-                            tile.setIcon(Icon.createWithResource(context, R.drawable.notification));
-                            tile.updateTile();
-                        }
-                    }));
-        } else {
-            Api.purgeIptables(context, true, new RootCommand()
-                    .setSuccessToast(R.string.toast_disabled)
-                    .setFailureToast(R.string.toast_error_disabling)
-                    .setReopenShell(true)
-                    .setCallback(new RootCommand.Callback() {
-                        public void cbFunc(RootCommand state) {
-                            Api.setEnabled(context, state.exitCode != 0, true);
-                            tile.setState(Tile.STATE_INACTIVE);// e() method form java, used in Kotlin as a property
-                            tile.setLabel(getString(R.string.inactive));
-                            tile.setIcon(Icon.createWithResource(context, R.drawable.notification_error));
-                            tile.updateTile();
-                        }
-                    }));
+                                tile.setState(Tile.STATE_ACTIVE);
+                                tile.setLabel(getString(R.string.active));
+                                tile.setIcon(Icon.createWithResource(context, R.drawable.notification));
+                                tile.updateTile();
+                            }
+                        }));
+            } else {
+                Api.purgeIptables(context, true, new RootCommand()
+                        .setSuccessToast(R.string.toast_disabled)
+                        .setFailureToast(R.string.toast_error_disabling)
+                        .setReopenShell(true)
+                        .setCallback(new RootCommand.Callback() {
+                            public void cbFunc(RootCommand state) {
+                                Api.setEnabled(context, state.exitCode != 0, true);
+                                tile.setState(Tile.STATE_INACTIVE);// e() method form java, used in Kotlin as a property
+                                tile.setLabel(getString(R.string.inactive));
+                                tile.setIcon(Icon.createWithResource(context, R.drawable.notification_error));
+                                tile.updateTile();
+                            }
+                        }));
+            }
         }
     }
 }
